moved enumes out
